% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{archs4_feature_info}
\alias{archs4_feature_info}
\title{Retrieves the feature (gene/transcript) information for the archs4 data}
\usage{
archs4_feature_info(feature_type = c("gene", "transcript"),
  source = c("human", "mouse"), distinct_symbol = TRUE)
}
\arguments{
\item{feature_type}{gene or transcript?}

\item{source}{human or mouse}
}
\value{
a tibble of information
}
\description{
Unfortunately only the gene symbols are stored in the gene-level count data
in \code{"meta/genes"}.
}
