% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/data-access.R
\name{file_info}
\alias{file_info}
\alias{archs4_file_info}
\title{Retrieves a table of files that back an Archs4Repository}
\usage{
file_info(x)

archs4_file_info(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{datadir}{the directory that stores the ARCHS4 data files}
}
\description{
A yaml iskept in the Archs4 data directory (\code{getOption("archs4.datadir")})
that links keys, (ie. \code{mouse_gene}) to the name of the file in the directory.
This abstraction is introduced so that the version of these files can be
updated with new downloads, and the user simply has to modify the yaml file
so that they are used downstream

Reference the "ARCHS4 Data Download" section in the vignette for more
information.
}
