% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/data-access.R
\name{file_path}
\alias{file_path}
\alias{archs4_file_path}
\title{Identify the file path on the system for specific ARCHS4 files.}
\usage{
file_path(x, key)

archs4_file_path(key, stop_if_missing = TRUE,
  file_info = archs4_file_info(datadir),
  datadir = getOption("archs4.datadir"))
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{key}{the lookup key for the file, ie. \code{"human_gene"} or \code{"mouse_gene"}.
The known keys are enumerated in \code{archs4_file_info()[["key"]]}.

[["key"]: R:[%22key%22}

\item{stop_if_missing}{defaults to \code{TRUE}, which causes this function to
throw an error if the file does not exist at the expected \code{file_path}.
Set this to \code{FALSE} to simply raise a warning}

\item{datadir}{the directory that stores the ARCHS4 data files}
}
\value{
a named (by \code{key}) character vector of paths to the filesystem that
correspond to the entries in \code{key}.
}
\description{
By default, this function will throw an error if a file does not exist
upon lookup. To return the \emph{expected} path to the, even if it does not
exist on the file system, set \code{stop_if_missing = FALSE}.
}
