% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-directory.R
\name{archs4_local_data_dir_create}
\alias{archs4_local_data_dir_create}
\title{Initialize a local datadir to act as an ARCHS4 data datadir}
\usage{
archs4_local_data_dir_create(datadir, stop_if_exists = TRUE)
}
\arguments{
\item{datadir}{the path to the datadir to create (or initialize) as an
ARCHS4 data datadir.}

\item{stop_if_exists}{by default, this function will \code{stop} if \code{datadir}
already exists. Set this to \code{FALSE} to continue. Setting it to \code{FALSE} is
convenient to initialize the target \code{datadir} with a \code{meta.yaml} file.
If a \code{meta.yaml} file already exists in \code{datadir}, then this function
will stop unconditionally. Move the \code{datadir/meta.yaml} out of the way
if you simply want to refresh it with the default version.}
}
\value{
invisibly returns the path to the \code{meta.yaml} in the target
\code{datadir}
}
\description{
Initialize a local datadir to act as an ARCHS4 data datadir
}
\details{
A local datadir needs to be created and initialized (wth a \code{meta.yaml}
file), to house ARCHS4 data for use in an Archs4Repository. This function
creates that datadir and copies an initial \code{meta.yaml} file.

Please refer to the vignette section "ARCHS4 Data Download" for more
details.
}
