% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-directory.R
\name{archs4_local_data_dir_validate}
\alias{archs4_local_data_dir_validate}
\title{Check "the health" of a local ARCHS4 data datadir}
\usage{
archs4_local_data_dir_validate(echo = TRUE,
  datadir = getOption("archs4.datadir"))
}
\arguments{
\item{echo}{echo validation diagnostic message to stdout via \code{\link[base:cat]{base::cat()}}}

\item{datadir}{the path to the datadir that stores local ARCHS4 data.
Defaults to \code{getOption("archs4.datadir")}.}
}
\value{
A string that indicates "what's wrong", or \code{TRUE} if validation
succeeds.
}
\description{
This function will notify the suer which files are missing from the
ARCHS4 data datadir, and what course of action they can use to
fix it.
}
