% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/data-access.R
\name{sample_covariates}
\alias{sample_covariates}
\alias{archs4_sample_covariates}
\title{Retrieves tibble of sample-level covariates available in mouse and human data.}
\usage{
sample_covariates(x, ...)

archs4_sample_covariates(datadir = getOption("archs4.datadir"), ...)
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{datadir}{the directory that holds the archs4 data}
}
\description{
Enumerate the sample covariates available in mouse and human data.
Note that the covariates available in human and mouse are the same
between the gene and transcript level files
}
