% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{archs4_sample_table}
\alias{archs4_sample_table}
\title{Lists the GEO series and samples available in the human and mouse datasets}
\usage{
archs4_sample_table()
}
\value{
a tibble of series_id, sample_id, species columns
}
\description{
This function queries the human and mouse gene expression matrices from
the ARCHS4 data release and combines their GEO series and sample identifiers
into a long table, annotated with the organism the samples come from.
}
\details{
This function executes very quickly (less thatn 0.10th of a second), so most
sample-level query functions in this package which you would think would
benefit from specifying human/mouse don't have to, as they will join into
this table to find out what species is being queried.
}
