% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/archs4-functional.R
\name{series_status}
\alias{series_status}
\alias{archs4_series_status}
\title{Checks which samples from a single series are present/absent.}
\usage{
series_status(x, id, ...)

archs4_series_status(id, sample_table = archs4_sample_table(datadir =
  datadir), datadir = getOption("archs4.datadir"), ...)
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{id}{a single GEO series id, ie. \code{"GSEnnnnn"}}

\item{sample_table}{the output from \code{\link[=archs4_sample_table]{archs4_sample_table()}}, which lists
the series_id,sample_id combinations found in the ARCHS4 repository.}

\item{datadir}{the directory that holds the archs4 data}
}
\value{
a tibble of information for a series.
}
\description{
Due to download or alignment issues, the ARCHS4 data processing pipeline may
not include all of the samples included in a particular GEO series. This
function will return a table with an \code{in_archs4} columns that indicates
whether or not a sample from a particular series is present in ARCHS4.
}
\examples{
info <- archs4_series_status("GSE89189")
}
