% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/archs4-functional.R
\name{sources}
\alias{sources}
\alias{archs4_sources}
\title{Lists the different sources ARCHS4 is built for}
\usage{
sources(x)

archs4_sources(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{datadir}{the directory that holds the archs4 data}
}
\value{
a character vector listing the different sources (organisms) that
the ARCHS4 repository has data for.
}
\description{
We hardocde these values in a lot of places ... who knows if one day these
are updated?
}
