% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{Archs4Repository}
\alias{Archs4Repository}
\title{An interface to a locally downloaded ARCHS4 dataset}
\usage{
Archs4Repository(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{datadir}{The directory that stores the ARCHS4 data.}
}
\value{
an Arhcs4DataSet object
}
\description{
This instantiates an object that acts as a central broker to handle queries
against the ARCHS4 dataset. You will need to download all of the files
enumerated in \code{archs4:::archs4.files()} and save them in \code{datadir}.
}
