% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist.R
\name{blacklist}
\alias{blacklist}
\title{Return a list of series/samples you may want to ignore for now}
\usage{
blacklist()
}
\description{
The entries that appear here so far are because the data appear to come from
single-cell experiments
}
