% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archs4-features.R
\name{create_augmented_feature_info}
\alias{create_augmented_feature_info}
\title{Create meta information for the genes and transcripts in the ARCHS4 dataset.}
\usage{
create_augmented_feature_info(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{datadir}{The directory that has the mouse and human expression
hdf5 files. There will be \code{SPECIES_FEATURETYPE_augmented_info.csv.gz} files
saved in this directory whe this function completes.}
}
\description{
This function creates \emph{all} of the feature-level CSV files for the features
enumerated in the \code{meta/genes} gene-level hdf5 file, and the
\code{meta/transcript} transcript identfiers in the transctipt-level hdf5 file for
the mouse and human files found in \code{datadir}.

\strong{In order for this to work} you have to download the approprate human and
mouse gtf files from ensembl and save them in \code{datadir}. Reference the
\code{\link[=archs4_local_data_dir_validate]{archs4_local_data_dir_validate()}} function.

For the initial relesae of the ARCHS4 dataset, the
\code{Homo_sapiens.GRCh38.90.gtf.gz} and \code{Mus_musculus.GRCm38.90.gtf.gz} were
used.
}
\details{
This function will write the augmented transcript- and gene-level files in
the \code{datadir}, using the following pattern:
\code{<organism>_<feature_type>_augmented_info.csv.gz}

Gene symbols are the only piece of information provided for the row-level
identifieres for the gene count matrices. Furthermore, the gene symbol used
in mouse are in all uppercase, which is not how genes are referred to there.
In order to augment the gene symbol information with gene-level identifiers
and other information, we parse relatively recent GTFs provided by GENCODE.

The fruits of the labor generated by this function are used by the
\code{\link[=archs4_feature_info]{archs4_feature_info()}} function.

Note that this function will replace already existing "\code{augmented}" files
if the already exist in \code{datadir}.
}
\seealso{
\code{\link[=archs4_feature_info]{archs4_feature_info()}}
}
