% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{create_augmented_gene_info}
\alias{create_augmented_gene_info}
\title{Create meta information for the genes and transcripts in the ARCHS4 dataset.}
\usage{
create_augmented_gene_info(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{datadir}{The directory that has the mouse and human expression
hdf5 files. There will be \code{SPECIES_FEATURETYPE_augmented_info.csv.gz} files
saved in this directory whe this function completes.}
}
\description{
This function creates feature-level metadata for the features enumerated in
the \code{meta/genes} gene-level hdf5 file, and the \code{meta/transcript} transcript
identfiers in the transctipt-level hdf5 file for the mouse and human files
found in \code{datadir}.
}
\details{
Gene symbols are the only piece of information provided for the row-level
identifieres for the gene count matrices. Furthermore, the gene symbol used
in mouse are in all uppercase, which is not how genes are referred to there.
In order to augment the gene symbol information with gene-level identifiers
and other information, we parse relatively recent GTFs provided by GENCODE.

The fruits of the labor generated by this function are used by the
\code{\link[=archs4_feature_info]{archs4_feature_info()}} function.
}
\seealso{
\code{\link[=archs4_feature_info]{archs4_feature_info()}}
}
