% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{datadir}
\alias{datadir}
\title{Retrieves the directory that contains the data for the Archs4Repository}
\usage{
datadir(x, ...)
}
\arguments{
\item{x}{an \code{Archs4Repository}}
}
\description{
Retrieves the directory that contains the data for the Archs4Repository
}
