% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5utils.R
\name{.h5read}
\alias{.h5read}
\title{Read data from an HDF5 file "with caution"}
\usage{
.h5read(file, name, index = NULL, start = NULL, stride = NULL,
  block = NULL, count = NULL, compoundAsDataFrame = TRUE,
  callGeneric = TRUE, read.attributes = FALSE, drop = FALSE, ...,
  native = FALSE, default_value = NA, default_dim = 1L)
}
\description{
This is a simple wrapper to \code{\link[rhdf5:h5read]{rhdf5::h5read()}} which returns a default value
if the "data element" (specified by \code{name}) does not exists within \code{file}.
We use this file to read from ARCHS4 hdf5 files when we want to provide a
little insurance to the evoling nature of their data formats.

For instance this function is used when we try to read somethign like
\code{"meta/reads_aligned"} because this information was not provided in earlier
versions of these datasets, however \code{"meta/genes"} may use \code{\link[rhdf5:h5read]{rhdf5::h5read()}}
directly because this has been around since "the beginning"
}
