% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{estimate_repository_norm_factors}
\alias{estimate_repository_norm_factors}
\title{Estimate normalization factors for datasets in the Archs4Repository}
\usage{
estimate_repository_norm_factors(a4)
}
\arguments{
\item{a4}{The \code{Arcsh4Repository}}
}
\value{
Invisibly returns a tibble of the the library size and normalization
factors for the expression data in \code{a4} (invisibly).
}
\description{
This function will serialize the results of the library size and
normalization factors into files inside \code{datadir(a4)}
}
