% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{feature_lookup}
\alias{feature_lookup}
\title{Perform a loose/fuzzy lookup for a gene/transcript feature.}
\usage{
feature_lookup(x, query, feature_type = "gene", source = "human", ...)
}
\arguments{
\item{x}{An Archs4Repository}

\item{query}{a character string of feature names to look for}

\item{feature_type}{"gene" or "transcript"}

\item{source}{organism dataset to lookup}
}
\value{
a tibble of features that match against the query. The first column
is the value of the query itself. If no match is found for a query, its
row is all \code{NA}.
}
\description{
This funciton facilitates exploratory data analyses by trying to find gene
or transcripts by different type of identifiers (symbol, ensembl_id, etc).
}
\examples{
a4 <- Archs4Repository()
features <- feature_lookup(a4, c("CFAP65", "PECR", "ENSG00000131408"),
                           feature_type = "gene", source ="human")
}
