% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{feature_source}
\alias{feature_source}
\title{Retrieves (organism) source of gene/transcript identifiers}
\usage{
feature_source(x, ...)
}
\arguments{
\item{x}{A vector of identifiers}
}
\description{
Parses ensembl identifiers and determines if they are for genes, or
transcripts, as well as the organism they should belong to (human, mouse)
}
