% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{libstats}
\alias{libstats}
\title{Extract the read depth and normalization factors for the samples}
\usage{
libstats(x, with_a4libsize = FALSE)
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{with_a4libsize}{If \code{TRUE}, includes an \code{a4libsize} column, which
was extracted from the \code{meta/reads_aligned} hdf5 file. Defaults to
\code{FALSE}.}
}
\value{
a tibble with sample_id, a4libsize, libsize, normfactor
}
\description{
Extract the read depth and normalization factors for the samples
}
