% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-utils.R
\name{lookup_gse}
\alias{lookup_gse}
\title{Retrieve information about a GEO series}
\usage{
lookup_gse(accession, fields = c("Accession", "Title", "Summary",
  "Overall-Design", "Type", "Pubmed-ID", "Sample"), ...)
}
\arguments{
\item{accession}{Scalar character, GEO series identifier e.g. GSE109171}

\item{fields}{Character vector specifying which fields to extract from the
XML file returned by GEO}

\item{...}{Additional arguments passed on to the \code{query_geo} function.}
}
\value{
List the requested \code{fields}
}
\description{
Queries NCBI GEO's REST interface to retrieve e.g. title, summary and the
list of samples for a GEO series.
}
\examples{
if (interactive()) {
  lookup_gse("GSE109171")
}
}
