% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-utils.R
\name{query_geo}
\alias{query_geo}
\title{Query NCBI GEO through its REST interface}
\source{
https://www.ncbi.nlm.nih.gov/geo/info/download.html
}
\usage{
query_geo(accession, target = c("self", "gsm", "gpl", "gse", "all"),
  validate = FALSE, verbose = FALSE)
}
\arguments{
\item{accession}{Scalar character, GEO identifier for a series (GSE), a
sample (GSM) or a platform (GPL).}

\item{validate}{Scalar boolean, validate the retrieved xml file against
NCBI's schema?}
}
\value{
xml2::xml_document object
}
\description{
Query NCBI GEO through its REST interface
}
\examples{
query_geo("GSE109171")
}
