% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-utils.R
\name{retrieve_sra_metadata}
\alias{retrieve_sra_metadata}
\title{Retrieve metadata for an SRA accession}
\usage{
retrieve_sra_metadata(x, from = c("ena", "ncbi"))
}
\arguments{
\item{x}{SRA identifier}

\item{from}{Scalar character, specifying either \code{ncbi} or \code{ena} as
the source database}
}
\value{
A tbl_df data.frame
}
\description{
This function uses the EBI's or the NCBI's REST APIs to retrieve information
about SRA data.
Study accessions (ERP, SRP, DRP, PRJ prefixes), experiment accessions
(ERX, SRX, DRX prefixes), sample accessions (ERS, SRS, DRS, SAM prefixes)
and run accessions (ERR, SRR, DRR prefixes) can be supplied.
For more information see \url{http://www.ebi.ac.uk/ena/browse/file-reports}
}
\note{
The output data.frame will be different for the two source databases.
}
\examples{
if (interactive()) {
   # retrieve study annotations
   retrieve_sra_metadata("SRP066489")
   # paired-end samples
   retrieve_sra_metadata("PRJEB2054", "ena") \%>\%
   dplyr::filter(sample_accession == "SAMEA728920")
}
}
