% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/data-access.R
\name{feature_info}
\alias{feature_info}
\alias{archs4_feature_info}
\title{Retrieves the feature (gene/transcript) information for the archs4 data}
\usage{
feature_info(x, feature_type = c("gene", "transcript"), source = c("human",
  "mouse"), distinct_symbol = TRUE, ...)

archs4_feature_info(feature_type = c("gene", "transcript"),
  source = c("human", "mouse"), distinct_symbol = TRUE,
  datadir = getOption("archs4.datadir"))
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{feature_type}{gene or transcript?}

\item{source}{human or mouse}
}
\value{
a tibble of information
}
\description{
Only the gene symbols (\code{meta/genes} in gene expression hd5 file) or entrez
transcript identifiers (\code{meta/ensemblid} for the transcript hdf5 file) are
stored in thse data. We use \code{\link[=create_augmented_gene_info]{create_augmented_gene_info()}} function to
generate and store extra metadata for these features, which are then appended
to these identifiers with this function.
}
\seealso{
\code{\link[=create_augmented_gene_info]{create_augmented_gene_info()}}
}
