% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/archs4-functional.R
\name{feature_info}
\alias{feature_info}
\alias{archs4_feature_info}
\title{Retrieves the feature (gene/transcript) information for the archs4 data}
\usage{
feature_info(x, feature_type = "gene", source = "human",
  augmented = TRUE, ...)

archs4_feature_info(feature_type = "gene", source = "human",
  augmented = TRUE, datadir = getOption("archs4.datadir"), ...)
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{feature_type}{gene or transcript?}

\item{source}{human or mouse}

\item{augmented}{include extra gene- or transcript-level features?
Default: \code{TRUE}}

\item{...}{pass through}

\item{datadir}{the directory that stores the ARCHS4 data files}
}
\value{
a tibble of information
}
\description{
Only the gene symbols (\code{meta/genes} in gene expression hd5 file) or entrez
transcript identifiers (\code{meta/transcript} for the transcript hdf5 file) are
stored in thse data. We use \code{\link[=create_augmented_feature_info]{create_augmented_feature_info()}} function to
generate and store extra metadata for these features, which are then appended
to these identifiers with this function.
}
\seealso{
\code{\link[=create_augmented_feature_info]{create_augmented_feature_info()}}
}
