% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{archs4_file_path}
\alias{archs4_file_path}
\title{Identify the file path on the system for specific ARCHS4 files.}
\usage{
archs4_file_path(what, datadir = getOption("archs4.datadir"))
}
\arguments{
\item{what}{what type of file? \code{"human_gene"} or \code{"mouse_gene"}}
}
\value{
the path on the filesystem for the path asked for
}
\description{
Identify the file path on the system for specific ARCHS4 files.
}
