% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/archs4-functional.R
\name{meta}
\alias{meta}
\alias{archs4_meta}
\title{Retrieves the meta information associated with an ARCHS4 datadir}
\usage{
meta(x)

archs4_meta(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{datadir}{the directory that holds the archs4 data}
}
\value{
a list-representation of the \code{meta.yaml} file
}
\description{
Retrieves the meta information associated with an ARCHS4 datadir
}
