% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R, R/archs4-functional.R
\name{sample_table}
\alias{sample_table}
\alias{archs4_sample_table}
\title{Lists the GEO series and samples available in the human and mouse datasets}
\usage{
sample_table(x, feature_type = c("gene", "transcript"), ...)

archs4_sample_table(feature_type = c("all", "gene", "transcript"),
  unroll_series = TRUE, datadir = getOption("archs4.datadir"))
}
\arguments{
\item{x}{an \code{Archs4Repository}}

\item{feature_type}{currently, the \code{"gene"} and \code{"transcript"} datasets
are not the same.}

\item{unroll_series}{There are some malformed series identifiers, like
\code{"GSE36025Xx-xXGSE49417Xx-xXGSE49847"} when the same sample_id appears
in multiple series. When this is \code{TRUE} (default), the series_id's are
unrolled and cleaned up.}

\item{datadir}{the directory that holds the archs4 data}
}
\value{
a tibble of series_id, sample_id, species columns
}
\description{
This function queries the human and mouse gene expression matrices from
the ARCHS4 data release and combines their GEO series and sample identifiers
into a long table, annotated with the organism the samples come from.
}
\details{
This function executes very quickly (less thatn 0.10th of a second), so most
sample-level query functions in this package which you would think would
benefit from specifying human/mouse don't have to, as they will join into
this table to find out what species is being queried.
}
