% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{archs4_sources}
\alias{archs4_sources}
\title{Lists the different sources ARCHS4 is built for}
\usage{
archs4_sources()
}
\value{
a character vector listing the different sources (organisms) that
the ARCHS4 repository has data for.
}
\description{
We hardocde these values in a lot of places ... who knows if one day these
are updated?
}
