% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{Archs4Repository}
\alias{Archs4Repository}
\title{An interface to a locally downloaded ARCHS4 dataset}
\usage{
Archs4Repository(datadir = getOption("archs4.datadir"))
}
\arguments{
\item{datadir}{The directory that stores the ARCHS4 data.}
}
\value{
an Arhcs4DataSet object
}
\description{
This instantiates an object that acts as a central broker to handle queries
against the ARCHS4 dataset. Please refer to the vignette for instructions
on how to setup a local directory to act as an Archs4Repository.
}
