% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-containers.R
\name{as.DGEList}
\alias{as.DGEList}
\title{Create a DGEList for the expression data of a series or set of samples.}
\usage{
as.DGEList(x, id, sample_columns = c("Sample_title",
  "Sample_source_name_ch1"), feature_type = c("gene", "transcript"),
  row_id = c("ensembl", "symbol"))
}
\arguments{
\item{id}{a vector of series or sample id's.}

\item{sample_columns}{the names of the sample covariates that are stored
in the ARCHS4 Dataset; a complete list of what covariates are available
in the ARCHS4 dataset is found using the \code{\link[=archs4_sample_covariates]{archs4_sample_covariates()}}
function.}

\item{feature_type}{do you want \code{"gene"} or \code{"transcript"} level expression?}

\item{row_id}{either \code{"ensembl"} or \code{"symbol"}. If this is \code{"ensembl"} and
\code{feature_type == "transcript"}, then we remove the rows from the count
dataset that we couldn't map symbol -> ensembl_gene_id for.}
}
\value{
a \code{DGEList} of results
}
\description{
Create a DGEList for the expression data of a series or set of samples.
}
\examples{
a4 <- Archs4Repository()
y <- as.DGEList(a4, "GSE89189", feature_type = "gene")
}
