% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archs4Repository.R
\name{feature_lookup}
\alias{feature_lookup}
\title{Perform a loose/fuzzy lookup for a feature}
\usage{
feature_lookup(x, query, feature_type = "gene", source = "human", ...)
}
\arguments{
\item{x}{An Archs4Repository}

\item{query}{a character string of feature names to look for}

\item{feature_type}{"gene" or "transcript"}

\item{source}{organism dataset to lookup}
}
\value{
a tibble of features that match against the query. If no match is
found for a query, its row is NA
}
\description{
Perform a loose/fuzzy lookup for a feature
}
